+++ src/Rindow/OpenBLAS/Math.c
@@ -3862,10 +3862,10 @@ ZEND_END_ARG_INFO()
 
 ZEND_BEGIN_ARG_INFO_EX(ai_Math_searchsorted, 0, 0, 12)
     ZEND_ARG_INFO(0, m)
+    ZEND_ARG_INFO(0, n)
     ZEND_ARG_OBJ_INFO(0, a, Interop\\Polite\\Math\\Matrix\\LinearBuffer, 0)
     ZEND_ARG_INFO(0, offsetA)
     ZEND_ARG_INFO(0, incA)
-    ZEND_ARG_INFO(0, n)
     ZEND_ARG_OBJ_INFO(0, x, Interop\\Polite\\Math\\Matrix\\LinearBuffer, 0)
     ZEND_ARG_INFO(0, offsetX)
     ZEND_ARG_INFO(0, incX)
