+++ testPHP/NDArrayPhp7.php
@@ -1,0 +1,10 @@
+<?php
+namespace RindowTest\OpenBLAS;
+
+class NDArrayPhp extends AbstractNDArrayPhp
+{
+    public function offsetGet( $offset )
+    {
+        return $this->doOffsetGet( $offset );
+    }
+}
