+++ testPHP/NDArrayPhp8.php
@@ -1,0 +1,12 @@
+<?php
+namespace RindowTest\OpenBLAS;
+
+use RindowTest\OpenBLAS\AbstractNDArrayPhp;
+
+class NDArrayPhp extends AbstractNDArrayPhp
+{
+    public function offsetGet( $offset ) : mixed
+    {
+        return $this->doOffsetGet( $offset );
+    }
+}
